requirements.txt
yfinance
kafka-python
newsapi-python
confluent-kafka